renamed rpn::Interpreter to rpn::Interp;  nested loops and progn fixes
Local vars are now contained both in a Progn and in
rpn::Interp::Privates.  When a Progn is evaluated, it's _locals are
copied in to Private's _locals.  When the Progn exits, the _locals are
removed from Private's _locals.

This does open the possiblity of recursive Progns having local
var collision.

: word1 FOR i ... NEXT ;

: word2 FOR i ... 0 9 word1 NEXT ;

After word1 is evaluated the local var 'i' is purged from Private's _locals.