make Interp run in a worker thread
- main interface to "eval()" and "parseFile()" are now async methods
with completionHandlers.

- Qt UI is converted to use the async mechanism.

Usage Pattern Intent is that UI thread will send commands and stuff
over to the Interp thread.  Interp thread will invoke a callback when
it's done with a request.

The thread safety is not generalized to an arbitrary number of threads
sending requests into the Interp, or accessing the Stack at any time.

To Do:
- Catch2 tests are not updated
- Interp.Stack is currently exposed and un-locked.  It may be ok due
to usage pattern(s).